refactor: modularize utility functions
