+++ src/tnfr/metrics/coherence.py
@@ -9,7 +9,8 @@ from typing import Dict
 from ..constants import ALIAS_THETA, ALIAS_EPI, ALIAS_VF, ALIAS_SI, COHERENCE
 from ..callback_utils import register_callback
 from ..glyph_history import ensure_history, append_metric
-from ..helpers import get_attr, clamp01
+from ..alias import get_attr
+from ..helpers import clamp01
 
 
 def _norm01(x, lo, hi):
