+++ src/tnfr/sense.py
@@ -7,10 +7,8 @@ import math
 import networkx as nx
 
 from .constants import ALIAS_SI, ALIAS_EPI, SIGMA
-from .helpers import (
-    get_attr,
-    clamp01,
-)
+from .alias import get_attr
+from .helpers import clamp01
 from .callback_utils import register_callback
 from .glyph_history import ensure_history, last_glyph, count_glyphs, append_metric
 from .constants_glyphs import (
