+++ tests/test_alias_sequence.py
@@ -1,7 +1,7 @@
 """Pruebas de helpers de alias con secuencias genéricas."""
 
 import pytest
-from tnfr.helpers import alias_get, alias_set
+from tnfr.alias import alias_get, alias_set
 
 
 def test_alias_get_accepts_sequence():
