+++ tests/test_validators.py
@@ -9,7 +9,8 @@ from tnfr.constants import (
     ALIAS_VF,
 )
 from tnfr.validators import run_validators
-from tnfr.helpers import set_attr_str, set_attr, read_structured_file
+from tnfr.alias import set_attr_str, set_attr
+from tnfr.io import read_structured_file
 from tnfr.config import load_config
 
 try:  # pragma: no cover - compatibilidad Python
